Add Podcasts to seed
These are the ten most recommended podcasts from the overcast.fm
app.

The idea is to add a Podcast title and the RSS feed link, so later we
can automatically get more information about the podcasts and its
episodes from the RSS feed.

The database is currently set up to require an itunes_image for each
podcast record, so we supply a placeholder string for now.