Write simple crawler to pull podcast details from RSS feed
The crawler takes a podcast and goes through its RSS feed. The details
of the channel are molded into a hash of attributes that are passed to
the podcast to update it.

The testing relies on the VCR gem which records the request to two RSS
feeds. This has the advantage that the test is fairly realistic and we
don't have to stub at a higher level, e.g. the `feed.channel`.

We use the Joe Rogan feed to get an example where the content is
explicit, so we can test the conditional that translates strings from
the RSS to booleans, i.e. "yes" to `true` and
"no" to `false`. The downside to this test is that it actually hits the
database. In the first test we just assert that `update` is being called
with the right attributes. Here however we only care about one specific
item.