Add a component to render each search result
The icon is a SVG which had to be loaded by Webpacker.[1]
The image is included as an object icon as recommended [2]. We could
also inline it, but since it's used repetitively it is best it benefits
from caching which only happens it is an image file the server fetches.
It does however fetch quite a few instances of this image
asynchronously in the first few fetches.

The component is based on the example GithubMenuItem component from
React-Bootstrap-Typeahead[3]

[1]
https://stackoverflow.com/questions/46844499/how-to-handle-images-in-a-rails-webpacker-react-app
[2] https://svgontheweb.com/#implementation
[3]
https://github.com/ericgio/react-bootstrap-typeahead/blob/master/example/components/GithubMenuItem.react.js