Fix Typeaheadsearch component
This commit does a number of things:

* It adds a search icon to be used with Bootstrap to add a search button
and demonstrate it's a search field. The button still doesn't link to
anything and needs to be hooked up.

* Also the filterBy option is set. This was necessary to include
description as the field the result is matched against and not just
title. The search is already performed against Elasticsearch, so
actually we just want to render the 10 results ES returned and not
filter further. The example "Nick Gillespie" would return 10 results
from Elasticsearch, but then not return results on the page, because the
name only appeared in the description, not in the title.

* This is still hardcoded to the localhost and should be passed the path
as a prop to work in different environmnets.