Allow testing with Elasticsearch
This adds before and after hooks to Rspec, so we can run elasticsearch
in tests.

They will run on a different port (9250) and create start nodes, create
indexes, then afterwards delete them and close the nodes.

This also requires a change to the configuration of VCR, to allow
requests to localhost, because Elasticsearch will make these requests.

Run specs with `elasticsearch: true` to run the above.

It would make sense to allow stubbing out of Elasticsearch in cases
where we don't want to use it and don't care about the callbacks that
index new writes to the episodes table.

https://medium.com/@rowanoulton/testing-elasticsearch-in-rails-22a3296d989
https://bonsai.io/blog/testing-elasticsearch-ruby-gems.html