Stub out Episode search module for each test
In tests were we don't want to use Elasticsearch, we can overwrite the
entire module, so the call back hooks to index new episodes are not
executred. This can be overwritten by assigning `elasticsearch: true`

The downside to this approach is that by default Elasticsearch indexing
will not happen in tests as it does in the codebase. That means
Elasticearch is deactivated by default in tests. If we want to use
it we have to explicitly state so.

On the plus side it allows us to write tests involving the episode model
and writing and updating episodes without triggering Elasticsearch out
of the box. This makes tests faster and easier to write.

Now we can remove `elasticsearch: true` from one of the specs.
Before the tests would fail, because we wrote episodes to the test DB
and it couldn't index them with Elasticsearch. So we activated
Elasticsearch and indexed all new episodes. However we didn't care about
this aspect of Elasticsearch in a test that is concerned with crawling
podcast episodes from a feed.

So with the new stubbing of Elasticsearch we can run the test without
having to use Elasticsearch at all.