Upgrade coffee-rails to version 5
This is to avoid the warning

```
DEPRECATION WARNING: Single arity template handlers are deprecated. Template handlers must
now accept two parameters, the view object and the source for the view object.
Change:
  >> Coffee::Rails::TemplateHandler.call(template)
To:
  >> Coffee::Rails::TemplateHandler.call(template, source)
```

The old version must have been a remnant of first installing Rails 5.
After the upgrade to Rails 6 the above error started appearing when running
specs.