Fix search feature spec
This commit makes two fixes to the feature spec:

* It adds a podcast before creating an episode for it. Previously the
  creation of the episode failed, because it needed a podcast record
  first.
* It adds a sleep of 1 second to the test. This was crucial to make it
  pass. Time needed to pass between making the search request by
  entering "Nick Gillespie" and clicking the search button that would
  make results appear on the page.

It took a long time to debug this, because I thought the issue might
rely on the test set up of Elasticsearch and the premature deletion of
the index. Also I expected that Webmock was blocking further requests to
localhost which were necesary for Elasticsearch. That seemed out to be
not relevant and localhost traffic was already allowed.

For debugging of the sleep issue it was helpful to be able to run the
spec in the browser and temporarily remove the headless option. That can
easily be toggled by changing

`Capybara.current_driver = :selenium_chrome_headless`
to
`Capybara.current_driver = :selenium_chrome`