Track pids with git for puma server
Rails 6 needs a process id (pid) file for the puma server.
This commit changes keeps ignoring most content in /tmp and also the
/tmp/pids directory, but adds a .keep file to allow pids files.

https://www.alanvardy.com/posts/38

Also I had to change the config in my global gitignore to allow this
file in tmp.

This article
https://www.atlassian.com/git/tutorials/saving-changes/gitignore proved
useful and in particular the command to debug what causes a file to be
gitignored or not: `$ git check-ignore -v /tmp/pids/.keep`