Remove rounded corners
This was primarily for the search result highlighting in the episode
view, but also applies to the <beta> tag on the bottom.

I expexted this to also work with a scss variable:
`$enable-rounded: false;` and documented here:
https://getbootstrap.com/docs/4.4/getting-started/theming/#sass-options
But that didn't end up working.

So I went with a solution mentioned here:
https://stackoverflow.com/questions/9742166/getting-rid-of-all-the-rounded-corners-in-twitter-bootstrap