Change Elasticsearch to search for phrase
When I search for "Nick Gillespie" I don't want matches for Nick, but
the whole phrase since it's about a mentions of a term or phrase, or
name.

This change makes it so Elasticsearch uses phrase search. Phrase search
has to be set as a sub-setting of multi_match. Otherwise you can't
search through multiple fields like title and description.

https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-match-query-phrase.html
https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-multi-match-query.html