Split into two types of search
The idea is to use typeahead_search to get immediate results and a
snappy search experience in the search box.

And then have a search endpoint that is hit with a normal http requst
and not xhr by hitting enter or clicking on the magnifying glass.

The main_search endpoint will be used later.