Require CSRF token for asynchronous search
Typeaahead async search returns with a json response to each query. That
could be abused as requests can be made from outside of the site.
Normally get requests are not restricted with CSRF, so it's not
necessary here.

But it's good to lock down these searches, so they can only be made from
within the app to avoid someone accessing the data from outside.

For this purpose we use the CSRF token that Rails generates and
authenticate each request with it.

Rails adds the token to the document. Since the request is made from JS
we read it from the DOM in React and then pass it as a header in the
request. In the controller we can use Rails methods to verify the
request has been made with the correct token.

The methods used in the controller come from
`actionpack/lib/action_controller/metal/request_forgery_protection.rb`

https://medium.com/rubyinside/a-deep-dive-into-csrf-protection-in-rails-19fa0a42c0ef

https://blog.eq8.eu/article/rails-api-authentication-with-spa-csrf-tokens.html

If the request is unauthorised the controller responds with a 403 and an
empty json response.

Just by including `valid_request_origin? &&
any_authenticity_token_valid?` in the controller the CSRF verification
is performed on all controller actions. Since we don't want it for the
get request to the main_search, that controller action needs to be
exempted.