Remove stubbing of the search module by default
Previously I thought that whenever a spec doens't explcitly run
elasticsearch, we would want to stub the entire searchable module, to
avoid any of the indexing code to be run for example.

But the empty module did cause trouble with verifying doubles. In the
searches_controller_spec I could not easily stub out
`Episode.typeahead_search`, because RSpec complained that Episode didn't
implement such a method.