Protect from forgery, with exception
If CSRF is not verified, it fails as it did before. But this time with
an exception.

This guards against accidentally executing code in the future despite an
invalid CSRF token.

See
https://nvisium.com/blog/2014/09/10/understanding-protectfromforgery.html
and the official docs:

https://api.rubyonrails.org/classes/ActionController/RequestForgeryProtection/ClassMethods.html#method-i-protect_from_forgery

It also makes a test in the searches controller spec easier to write.
Now, with the wrong CSRF token and `protect_from_forgery` disabled, an
`InvalidAuthenticityToken` is raised.