Install Sidekiq
This installs Sidekiq. Redis is automatically installed as well as the
datastore where jobs are enqueued.

It also adds a admin route to view sidekiq stats under /admin/sidekiq

Note that this route should be locked down with devise and is currently
publicly available. However production is still behind basic
authentication.

https://github.com/mperham/sidekiq/wiki/Getting-Started