Revert "Remove stubbing of the search module by default"
This reverts commit 0dea702e44aad7f58dcd47e468251f6d831cc4d0.

It turns out that removing this stubbing of the module made the
podcast_crawler_spec fail.

It fails, because the Elasticsearch callbacks try to index the newly
created episodes in Elasticsearch, but can't because the connection to
ES can't be established.

This reverts the change introduced in 0dea702e44aad7f. It turns out now,
that the stubbing of Episode.typeahead_search still works. The
verification of the existence of this method does no longer cause a
problem. In the meantime RSpec had been updated to 4.0 beta, so this
could have fixed the issue. The issue was that the stubbing of the
module removed `Episode.typeahead_search` and hence RSpec said we can't
stub that method since it is not actually implemented.

The reason of this failure was a bit hard to uncover as it was disguised
by the rescuing of StandardError, as described here [1].

[1] 94d994642ff0db578