Stub Elasticsearch indexing of Episodes
The specs in podcast_crawler_spec failed. Because after creating
episodes callbacks in the Episode model would index new episodes in
Elasticsearch. The network connection couldn't be established, so it
failed.

This commit stubs the indexing method in the Elasticsearch library which
is called when creating or modifying episodes [1].

Stubbing the Elasticsearch library might not be the best idea. An
alternative could be to stub out the whole module that is included in
the Episode module.

[1] ~/.gem/ruby/2.6.2/bundler/gems/elasticsearch-rails-7bbe7ec89e9a/elasticsearch-model/lib/elasticsearch/model/adapters/active_record.rb:65