Return integers rather than time duration objects
The Crawler::PodcastCrawler#convert_time_to_seconds method returned time
duration objects, which were stored in the database as integers and
hence were processed correctly. But since in the crawler we pass a
hash to `podcast.episodes.create` it makes more sense to pass normal
Ruby objects, such as integers and string.

This became apparent when writing a spec and stubbing out the objects.
It felt strange to create a hash in the spec with a time duration object.