Add spec to unit test parsing of episodes from RSS feed
Rather than using VCR this stubs out the entire http request and then
sets up an OpenStruct which is returned as a feed entry.

This allows us to test different scenarios of the feed structure in isolation.
The OpenStruct doesn't mirror the feed object exactly. But it imitates
it's object structure and the way entry elements can be accessed with
instance methods such as #description or #title.

Now we can set these to return different values or nil and use the test
to assert all scenarios are handled correctly.