Handle description not being present in RSS feed
There are RSS feeds such as Joe Rogan that don't have a description in
the itunes_summary node. That means that previously the description
field in our episode entry was left empty.

This change should now look for the description in the content and
summary sections as well if itunes_summary is left empty, so we are more
likely to retrieve a description. Description is one of the few fields
that is indexed in search and exposed multiple times throughout the
site, so important it's present wherever possible. If it is empty, it
can also not be searched.