Try to get episode description from different sources
If a feed entry doesn't have a description FeedJira would raise a
NoMethodError. Hence we now proceed with more caution and check first
which methods the entry object responds to before sending it that
method.

I investigated this with this RSS feed: "http://feeds.feedburner.com/paperclipping"
After this fix I saw that the parsing of these episodes still failed.
The RSS didn't have an enclosure URL. However it was broken in different
ways and the image part of the feed entry linked to an mp4 file.

Services like Overcast also couldn't properly process the image:

https://overcast.fm/itunes202043548/paperclipping-scrapbooking-videos

Furthermore, the feed's last entry is from 2010, so possibly it's ok for
this to not be processed properly, buy raise an error because it is a
feed that is only partially processable.