Add mailers queue to Sidekiq config
Emails from Rails triggered with `#deliver_later` will be sent to the
mailers queue in Sidekiq.

Alternative we could start Sidekiq with the `-q mailers` option. But it
seems easier to configure Sidekiq via the codebase and track changes in
git. That way we can just start sidekiq without configuration flags by
using `bundle exec sidekiq`.

Due to the order of the queues jobs in the default queue will only be
processed if mailers is empty. So notification emails are prioritised
over fetching new podcast episodes.

https://github.com/mperham/sidekiq/wiki/Advanced-Options#queues