Create search_terms
A user can have many search terms and a search_term belongs to a user.
Also a foreign key constraint is added. So if a user is deleted, the
search_terms should be removed as well. This should create a good
database integrity and not leave orphaned search_term records behind
shuold a user be removed.