Don't submit search term form as ajax request
By default `form_with` forms are set to `remote: true`.
This means they are submitted as an ajax request.

This caused the flash[:error] message from the controller to not be set
in the view. Therefor a form with errors would not submit successfully,
but not show any errors to the user.

This commit removes `remote: true` from the form. Hence it's submitted
as a normal HTTP request. Now the flash[:error] shows.

The downside is that the page refreshes upon form submission. This
provides a rougher user experience. However a successful post request
would be answered with a redirect to the root path anyhow, so the
experience mainly changes for unsuccessful form submissions.

An alternative would be to keep the form to `remote: true` and figure
out a better way of setting errors on the form. In part this is
explained here:

https://guides.rubyonrails.org/working_with_javascript_in_rails.html#form-with

See the following for setting `local: true`, to overwrite the remote
default.
https://api.rubyonrails.org/v6.0.2.2/classes/ActionView/Helpers/FormHelper.html#method-i-form_with