Add SearchTermMatches
A search term can have many search term matches.
A search term match belongs to a search term
A search term match belongs to an episode.

A search term match should also make reference to the field where it
matched.

I briefly considered modeling this as a has_many through: relation, but
this is not actually a many to many relationship. A search_term_match
should only ever belong to one search_term.