Add user_id to search_term_matches
A search_term_match should be unique to a term and an episode, but only
per user. That means that another user could have set the same search
term and it was found in the same episode, but since they are a
different user, the record should still be added to the
search_term_matches table.

By creating a relationship to users this uniqueness can now be added in
a commit later on.

With the presence change we can add an index which should be more
performant for checking for duplicates.

It's worth mentioning, that just for the purpose of finding the user
that a search_term_match belongs to one could already traverse from the
search_term_match to the search_term and then the user, but that seemed
inefficient at times of running queries and checking for a presence.