Ensure uniqueness of combination of search_term, episode and user
A user should only have one match of a search_term in an episode. But a
different user could have the same combination of search_term and
episode.

This might be somewhat unnecessary. A search_term is already unique to a
user, so that implicitly enforces that the same search_term isn't
prevented from being created because the same search_term#name was
already found on the same episode for another user.

But it feels like this is a sound step to ensuring data integrity even
further.