Add a search to crawl episodes for a search term
This query searches Elasticsearch for a search_term, but also takes a
date that ensures no result is older than the specified date.

The purpose of this filter is to automatically and regularly query the
ES index, but not dig up results that had been reported as a
search_term_result before.