Add SearchTermCrawlerService
This crawler searches Elasticsearch for a specific search term, but
ignores results before the passed in date. The results are stored in the
search_term_matches table.

Note that the formatted_results object which is a Hash with indifferent
access created by Elasticsearch can also access the id with an instance
method, rather than using `Hash#fetch`, i.e. `r.id`.
However that would also require the test _source object to not be a hash,
but something that can receive #id and since that seemed more involved
to set up, we went with Hash in the implementation.