Add SearchTermMatch#notified
When a user has been notified of a SearchTermMatch, notified receives
the Timestamp. This timestamp will show when a notification has been
sent and also if a notification has been sent.

This modeling of notifications of search_term_matches replaces the
previous attempt of modeling notifications as a new table. However
these merely had a sent_on column. They were related by a has_one
relationship to search_term_matches and absennce of a notification
record signalled to the search_term_match that the user had not been
notified. This was complicated by the fact that creating a new
notification record for a search_term_match did not seem to update the
foreign key notification_id on search_term_matches.