Add notification scopes
The purpose of the due_for_notification scope is to get all
search_term_matches that the user has not been notified of and still
should. This excludes search_term_matches belonging to a search_term
that has been created today.

There is no need to notify a user about a match that was found the same
day they set up their search_term as they will already know about these
matches.

On the other hand, the search_term_matches are already set up to not be
created with matches from the past. So we might remove the second part
of this scope here and notify of everything that has a match since the
search_term was created.

In addition to a spec this commit adds new factories to set up
search_terms and search_term_matches more easily.

The http calls to elasticsearch had to be stubbed out, because upon
creation of an episode, elasticsearch wants to index these.