Don't pass date to SearchTermCrawlerService
The date that is passed to the search query can be calculated
within the service.

The search for appearances of the search_term (search_term_match) should
return records no older than the last search_term_match. If there has
been no match yet, then start search with the date of today. That way we
don't show the user results from the past that matched.