Add feature spec to test crawling of episodes for search_term
This feature spec mirrors the app's behavior, without user involvement.
It traces the steps of
1.) The cronjob executes a rake task
2.) Rake task goes through search_terms and passes each to a worker
3.) The worker calls the SearchTermCrawlerService
4.) The Service uses Elasticsearch to search through all episodes for
    the search term
    Any matches are stored in the search_term_matches table

We need to wait for Sidekiq after the job has been enqueued before
draining the queue and thereby executing the job.

Unfortunately at this point the cronjob and the rake task are not used
in this feature spec, but the code that the rake task calls is executed
directly in the spec.