Add feature spec for notifying users of search term matches
This feature spec traces the following steps

1.) Cronjob calls rake task
2.) The rake task goes through all search_term_matches and creates a
    Sidekiq job for each
3.) The Sidekiq job calls the notifier service
4.) The service sends an email to the user to alert of the search term
    match and records that the user has been notified

The code inside the rake task is executed directly in the feature spec.
The cronjob and rake task are hence not really tested.

We need to wait for Sidekiq after the job has been enqueued before
draining the queue and thereby executing the job. Elasticsearch is
needed since the creation of a search term match needs the existense of
an episode which indexes it in Elasticsearch. The email that is sent
also needs the episode to exist.