Upgrade elasticsearch to version 7
Render runs 7.4 on production and with Rails we used version 6.
This led to an error around indexing
("The mapping definition cannot be nested under a type [_doc]").

The tests failed after this update, because elasticsearch-model/rails
would not keep elasticsearch-api on version 5. The explicit addition
updates the wider set of elasticsearch libraries to 7.

https://github.com/elastic/elasticsearch-rails/issues/877