Fetch episode description
Previously the episode description couldn't be fetched correctly in
cases like the Coronavirus podcast.

That is because it was checked whether the episode responded to methods
like itunes_summary. However when it responded, i.e. not failed with a
NoMethodError the field could still be nil. This is what happened here,
so it was not attempted to get the description from a different field.