Stub indexing in Elasticsearch
Whenever a spec is run without the meta tag :elasticsearch, we do not
want to index episodes that were created.

Indexing requires a http request to the Elasticsearch instance.
Previously we either stubbed out individual requests to the indexing
endpoint for docs for each episode or set the stubbing in invidiual
specs.

Now that the indexing is stubbed from the spec helper, we can remove it
from existing specs and don't have to worry about this part in the
setup.