Change SearchTermMatchNotifierService to handle one email for multiple search_term_matches
This refactors the service, so a user can receive one email with
multiple search terms.

The service receives a list of user_ids, as it will go through Sidekiq
which can only receive plain ruby objects. From the ids the users are
found and for each user the search_terms are retrieved, that have
search_term_matches that the user has not been notified about yet.

The queries run from user -> search_term -> search_term_match