Change feature spec to adjust to emailing only once about multiple search_term_matches
The indexing in Elasticsearch was created in 36594d6
and justified by episodes being created as a secondary effect of
creating search_term_matches. But the existence of an episode does not
mean it has to be indexed in Elasticsearch. And since Elasticsearch
functioanlity is not necessary for this feature spec, we can remove the
elasticsearch flag from the spec and rely on the stubbing of indexing
documents in Elasticsearch via the spec_helper before blocks.