Fix Timecop bug
This fixes a bug that came up, because TimeCcop would not reset the time
after it was initiated.

services/search_term_match_notifier_service_spec.rb failed only when run
with the whole test suit, but not when run in isolation. It took me a
while to realise it was because Rubocop sets the time to April 1st in
the Searchable spec, but doesn't reset it.

The error was for comparing dates

expected: "2020-04-01"
got: "2020-03-31"

The Rubocop docs show how to reset the time after Rubocop is no longer
used and this fixes the bug.

What also fixed the bug was using `Date.current.to_s`, since the
timezone of the app was used in both side of the comparison and not
`Date.today`, which uses the current system time, while the app runs on
UTC time.