Add sign in form using Devise
These views were automatically generated by using the command

$ rails generate devise:views users
https://github.com/heartcombo/devise#configuring-views

Most of the views are not needed and could be removed. Only

views/users/sessions/new.html.erb is actually used in that the layout is
slightly customised, mainly to center the form on the page.

This allows users to sign in and is currently particulary important to
authenticate as an admin and get access to /admin/sidekiq